Allow looged-in users to ignore permissions when accessing resource that belongs to them

+++ database/migrations/2016_10_16_200737_create_roles_table.php
@@ -15,7 +15,7 @@ class CreateRolesTable extends Migration
     {
         Schema::create('roles', function (Blueprint $table) {
             $table->increments('id');
-            $table->smallInteger('permissions_sum')->unsigned()->default(0);
+            $table->bigInteger('permissions_sum')->unsigned()->default(0);
             $table->string('name');
             $table->string('title')->nullable();
         });
