+++ database/migrations/2017_02_03_184300_create_permissions_table.php
@@ -15,7 +15,7 @@ class CreatePermissionsTable extends Migration
     {
         Schema::create('permissions', function (Blueprint $table) {
             $table->increments('id');
-            $table->smallInteger('bit')->unsigned()->default(1);
+            $table->bigInteger('bit')->unsigned()->default(1);
             $table->string('name')->unique();
             $table->string('title')->nullable();
         });
